requirements.txt
streamlit==1.32.0
pymongo==4.6.1
python-dotenv==1.0.1
pandas==2.2.1
